requirements.txt
Django==2.0.3
djangorestframework==3.8.2
drf-yasg==1.10.0
gunicorn==19.7.1
